requirements.txt
asgiref==3.3.1
Django==3.1.5
psycopg2==2.8.6
pytz==2020.5
sqlparse==0.4.1
djangorestframework==3.12.2
drf_social_oauth2==1.0.8
django-cors-headers==3.7.0